==> Spatial/visium_heart-master/snATAC_seq/data_integration/02_quality_check.ipynb <==
suppressMessages(library(ArchR))
suppressMessages(library(patchwork))
suppressMessages(library(ggplot2))
suppressMessages(library(dplyr))
suppressMessages(library(cowplot))
suppressMessages(library(ggrepel))

### Set parameters

set.seed(42)
addArchRThreads(threads = parallel::detectCores() - 2)
addArchRGenome("hg38")

proj <- loadArchRProject("../data/VisiumHeart", showLogo = FALSE)

# number of cells per sample
df <- proj@cellColData %>%
    as.data.frame() %>%
    group_by(Sample) %>%
    summarise(n_cells = n(),
             n_fragments = sum(nFrags))

p <- ggplot(data = df, aes(x = reorder(Sample, -n_cells), y = log10(n_cells))) +
    geom_bar(stat = "identity") +
    geom_text(aes(label=n_cells), position=position_dodge(width=0.9), vjust=-0.25) +
    xlab("") + ylab("Number of cells (log10)") +
    theme_cowplot() +
    theme(axis.text.x = element_text(angle = 60, hjust = 1))

options(repr.plot.width = 20, repr.plot.height = 5)

p

### Visualize quality control

## Visualize data quality
p1 <- plotGroups(ArchRProj = proj, 
                 groupBy = "Sample", 
                 colorBy = "cellColData", 
                 name = "TSSEnrichment",
                 alpha = 0.4,
                 plotAs = "violin",
                 addBoxPlot = TRUE)

p2 <- plotGroups(ArchRProj = proj, 
                 groupBy = "Sample", 
                 colorBy = "cellColData", 
                 name = "log10(nFrags)",
                 plotAs = "violin",
                 alpha = 0.4,
                 addBoxPlot = TRUE)

options(repr.plot.width = 10, repr.plot.height = 5)

p1
p2

## Session information

sessionInfo()
