==> Spatial/visium_heart-master/snATAC_seq/data_integration/03_create_matrix.ipynb <==
suppressMessages(library(ArchR))
suppressMessages(library(patchwork))
suppressMessages(library(ggplot2))
suppressMessages(library(dplyr))
suppressMessages(library(cowplot))
suppressMessages(library(ggrepel))

set.seed(42)
addArchRThreads(threads = parallel::detectCores() - 2)
addArchRGenome("hg38")

proj <- loadArchRProject("../data/VisiumHeart", showLogo = FALSE)

proj <- addGroupCoverages(ArchRProj = proj, groupBy = "Sample",
                          maxCells = 5000,
                          minCells = 5000,
                          force = TRUE)

## peak calling
pathToMacs2 <- findMacs2()

proj <- addReproduciblePeakSet(
    ArchRProj = proj, 
    groupBy = "Sample", 
    cutOff = 0.05,
    pathToMacs2 = pathToMacs2
)

proj <- addPeakMatrix(proj)

## save peak matrix
peakMatrix <- getMatrixFromProject(proj,
                                   useMatrix = "PeakMatrix")

counts <- peakMatrix@assays@data$PeakMatrix
df_rangers <- as.data.frame(peakMatrix@rowRanges@ranges)

rownames(counts) <- paste(peakMatrix@rowRanges@seqnames,
                          df_rangers$start,
                          df_rangers$end,
                          sep = "_") 

saveRDS(counts, file = "../data/VisiumHeart/PeakMatrix.Rds")

## save gene matrix
atac <- getMatrixFromProject(ArchRProj = proj,
                             useMatrix = "GeneScoreMatrix")

counts <- atac@assays@data$GeneScoreMatrix
rownames(counts) <- atac@elementMetadata$name

saveRDS(counts, file = "../data/VisiumHeart/GeneScoreMatrix.Rds")

saveArchRProject(ArchRProj = proj, 
                 load = FALSE)
sessionInfo()
