==> Spatial/visium_heart-master/snATAC_seq/data_integration/08_viz_marker_gene.ipynb <==
suppressMessages(library(ArchR))
suppressMessages(library(Seurat))
suppressMessages(library(Signac))
suppressMessages(library(cowplot))
suppressMessages(library(dplyr))
suppressMessages(library(tidyr))
suppressMessages(library(ComplexHeatmap))
suppressMessages(library(RColorBrewer))

obj.atac <- readRDS("../data/VisiumHeart/snATAC.annotated.Rds")
obj.atac

## add gene activity matrix
gene.activities <- readRDS("../data/VisiumHeart/GeneScoreMatrix.Rds")

gene.activities <- gene.activities[, colnames(obj.atac)]

obj.atac[['GeneActivity']] <- CreateAssayObject(counts = gene.activities)

DefaultAssay(obj.atac) <- "GeneActivity"

obj.atac <- obj.atac %>% 
        NormalizeData() %>%
        FindVariableFeatures() %>%
        ScaleData() %>%
        RunPCA(verbose = FALSE)

Idents(obj.atac) <- "cell_type"

all.markers <- FindAllMarkers(obj.atac, only.pos = TRUE, min.pct = 0.25, logfc.threshold = 0.25)

all.markers$cluster <- factor(all.markers$cluster, levels = c("CM", "Fib", "Endo", "Lymphoid", "Myeloid", "Pericyte", "vSMCs", "Neuronal"))

all.markers <- all.markers[order(all.markers$cluster), ]

df <- all.markers %>%
    group_by(cluster) %>%
    slice_max(n = 10, order_by = avg_log2FC)

## get average gene activity score
avg <- AverageExpression(
  obj.atac,
  assays = "GeneActivity",
  features = df$gene,
  return.seurat = FALSE,
  group.by = "cell_type",
  slot = "data",
  verbose = TRUE,
)


df_plot <- avg$GeneActivity

head(df_plot)

df_plot_scale <- t(apply(df_plot, 1, scale))
colnames(df_plot_scale) <- colnames(df_plot)
rownames(df_plot_scale) <- rownames(df_plot)

head(df_plot_scale)

df_plot_scale <- df_plot_scale[, c("CM", "Fib", "Endo", "Lymphoid", "Myeloid", "Pericyte", "vSMCs", "Neuronal")]

df_plot_scale <- df_plot_scale[rev(rownames(df_plot_scale)), ]

options(repr.plot.width = 4, repr.plot.height = 14)

p <- Heatmap(as.matrix(df_plot_scale),
             name = "Gene Activity",
             cluster_columns = FALSE,
             cluster_rows = FALSE,
             show_row_names = TRUE,
             rect_gp = gpar(col = "black", lwd = 0.5),
             col = ArchR::paletteContinuous()
            )

p

saveRDS(df_plot_scale, "../data/VisiumHeart/marker_heatmap.Rds")

## Session information
sessionInfo()
